Update follow buttons to the new UI
